voxel: Separate common cli
