+++ tools/rust-rofi_voxel_reconfig/src/bin/rofi-voxel_reconfig.rs
@@ -1,56 +1,7 @@
 use anyhow::Result;
 use clap::Parser;
-use input::Input;
 use rofi_voxel::reconfiguration;
-
-pub mod input {
-    use anyhow::{ensure, Context, Result};
-    use std::fs::File;
-    use std::io::{BufReader, Read};
-
-    #[derive(Debug, Clone)]
-    pub enum Input {
-        File(String),
-        StdIn,
-    }
-    impl Input {
-        pub fn from_arg<TString: AsRef<str> + Into<String>>(file_arg: TString) -> Self {
-            match file_arg.as_ref() {
-                "-" => Self::StdIn,
-                _ => Self::File(file_arg.into()),
-            }
-        }
-
-        pub fn get_reader(&self) -> Result<BufReader<Box<dyn std::io::Read>>> {
-            match self {
-                Input::File(path) => Ok(BufReader::new(Box::new(
-                    File::open(path).with_context(|| format!("Could not open file {path:?}"))?,
-                ))),
-                Input::StdIn => Ok(BufReader::new(Box::new(std::io::stdin()))),
-            }
-        }
-
-        pub fn read_all(&self) -> Result<String> {
-            let mut result = String::new();
-            self.get_reader()?.read_to_string(&mut result)?;
-            Ok(result)
-        }
-
-        pub fn check_max_one_stdin<'a, TInputs: IntoIterator<Item = &'a Self>>(
-            inputs: TInputs,
-        ) -> Result<()> {
-            ensure!(
-                inputs
-                    .into_iter()
-                    .filter(|&input| matches!(input, Self::StdIn))
-                    .count()
-                    <= 1,
-                "Cannot redirect multiple files to stdin"
-            );
-            Ok(())
-        }
-    }
-}
+use rofi_voxel_cli::FileInput;
 
 /// Compute RoFI reconfiguration from init to goal by using voxels
 #[derive(Debug, Parser)]
@@ -72,10 +23,10 @@ struct InputWorlds {
 
 impl Cli {
     pub fn get_worlds(&self) -> Result<InputWorlds> {
-        let init = Input::from_arg(&self.init_world_file);
-        let goal = Input::from_arg(&self.goal_world_file);
+        let init = FileInput::from_arg(&self.init_world_file);
+        let goal = FileInput::from_arg(&self.goal_world_file);
 
-        Input::check_max_one_stdin([&init, &goal])?;
+        FileInput::check_max_one_stdin([&init, &goal])?;
 
         let init = init.read_all()?;
         let goal = goal.read_all()?;
